backend/requirements.txt
django
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
python-dotenv
dj-database-url
PyJWT

# Type stubs (dev)
django-stubs
django-stubs-ext
djangorestframework-stubs
types-psycopg2
dj-rest-auth
django-allauth
